requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
jinja2==3.1.4
google-generativeai==0.8.3
pillow==10.4.0
python-multipart==0.0.12
python-dotenv==1.0.1
mistralai>=1.0.0
pyyaml>=6.0.1
aiofiles>=23.0.0
httpx>=0.25.0